+++ assets/javascripts/kitten/karl/navigation/horizontal-nav.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react'
 import { HorizontalNav } from 'kitten/components/navigation/horizontal-nav'
+import { Row } from 'kitten/components/grid/row'
 
 export const KarlHorizontalNav = props => {
   const horizontalNavItems = [
@@ -10,6 +11,29 @@ export const KarlHorizontalNav = props => {
   ]
 
   return (
-    <HorizontalNav items= { horizontalNavItems } />
+    <HorizontalNav items={ horizontalNavItems } />
+  )
+}
+
+export const KarlHorizontalNavBrowsing = props => {
+  const rowStyles = {
+    backgroundColor: '#f6f6f6',
+    borderTop: '1px solid #eee',
+  }
+
+  const horizontalNavItems = [
+    { text: 'Projet d’entreprise', href: '#', selected: true },
+    { text: 'Énergies renouvelables', href: '#' },
+    { text: 'Immobilier', href: '#' },
+  ]
+
+  return (
+    <Row style={ rowStyles }>
+      <HorizontalNav
+        items={ horizontalNavItems }
+        height="100px"
+        center
+      />
+    </Row>
   )
 }