+++ spec/dummy/client/entries/app-kitten.js
@@ -202,7 +202,8 @@ import {
 
 // Navigation
 import { HorizontalNav } from 'kitten/components/navigation/horizontal-nav'
-import { KarlHorizontalNav } from 'kitten/karl/navigation/horizontal-nav'
+import { KarlHorizontalNav,
+         KarlHorizontalNavBrowsing } from 'kitten/karl/navigation/horizontal-nav'
 import { TabBar } from 'kitten/components/navigation/tab-bar'
 
 // Notifications
@@ -445,8 +446,9 @@ ReactOnRails.register({
 
   // Navigation
   HorizontalNav,
-  KarlHorizontalNav,
   TabBar,
+  KarlHorizontalNav,
+  KarlHorizontalNavBrowsing,
 
   // Notifications
   Alert,